requirements.txt
Django==2.0.2
django-filter==1.1.0
django-leaflet==0.23.0
djangorestframework==3.7.7
djangorestframework-gis==0.12
django-rest-framework-social-oauth2==1.1.0
earthengine-api==0.1.135
geoip2==2.7.0
geopy==1.12.0
google-api-python-client==1.6.5
gunicorn==19.7.1
model_mommy==1.5.1
numpy==1.14.1
psycopg2==2.7.4
psycopg2-binary==2.7.4
pyOpenSSL==17.5.0
python-geohash==0.8.5
pytz==2018.3
scipy==1.0.0
